fix(clippy): remove unnecessary as u32 casts in initialize
In soroban-sdk, Vec::len() already returns u32, so the as u32
casts on accepted_assets.len() and milestone_count were redundant
and triggered clippy::unnecessary-cast warnings.

+++ campaign/src/lib.rs
@@ -166,8 +166,8 @@ impl CampaignContract {
             &creator,
             goal_amount,
             end_time,
-            accepted_assets.len() as u32,
-            milestone_count as u32,
+            accepted_assets.len(),
+            milestone_count,
             env.ledger().sequence(),
         );
